Rename fullscore_application_controller to app_controller
