+++ src/actions/set_current_gui_score_editor_action.cpp
@@ -4,7 +4,7 @@
 
 #include <fullscore/actions/set_current_gui_score_editor_action.h>
 
-#include <fullscore/fullscore_application_controller.h>
+#include <fullscore/app_controller.h>
 #include <fullscore/gui_score_editor.h>
